refactor(budget): relocate Budget to cube_harness/budget.py + rename turn→agent_step
Two coupled changes:

1. **Move `Budget` + `BudgetExceeded` from `tool.py` to new
   `cube_harness/budget.py`.** Budget is a cross-cutting concept
   (enforced by tool wrappers, fed by streamer, queried by agents); it
   doesn't belong in the tool module just because that's where
   MonitoredTool reads it. `tool.py` re-exports both names so existing
   call sites `from cube_harness.tool import Budget` keep working.

2. **Rename the per-agent-step counter for honesty.**
     - `max_turns` → `max_agent_steps`
     - `turns` → `agent_steps`
     - `bump_turn()` → `bump_agent_step()`
   `EventStreamer.on_step()` already used "step" vocabulary, and
   `summary_stats` exports `n_agent_steps` — Budget was the last
   place using "turns" (overloaded: one turn used to mean one LLM
   call; now it's one agent-loop iteration which may make 0..N LLM
   calls). The unambiguous name matches what's actually counted.

   `Budget.__str__` output reflects the rename: "budget used:
   agent_steps 34/150 (23%), …" instead of "turns 34/150 (23%)".

All call sites updated: src/ + tests/ + recipes/ + the agent docstrings.

922 unit tests pass; lint clean; agent_owns_loop_events +
genny_parallel_recorder smokes green.
