+++ src/cube_harness/agents/genny_parallel.py
@@ -106,8 +106,8 @@ class GennyParallel(Genny):
         obs = initial_obs
         while True:
             agent_output = await asyncio.to_thread(self.step, obs)
-            # Bump `budget.turns` once per agent step (mirror Agent.run).
-            # See Budget.bump_turn / EventStreamer.on_step for why
+            # Bump `budget.agent_steps` once per agent step (mirror Agent.run).
+            # See Budget.bump_agent_step / EventStreamer.on_step for why
             # turn-counting is per-step, not per-LLM-call.
             if self._recorder is not None:
                 self._recorder.on_step()
