+++ tests/test_budget_concurrency.py
@@ -21,7 +21,7 @@ def test_concurrent_bump_tool_calls_count_correctly() -> None:
     """Fire N concurrent bump_tool_calls() from a thread pool; assert
     the final counter equals N. Without the lock, the read-modify-write
     of tool_calls would race and undercount."""
-    budget = Budget(max_turns=100_000)
+    budget = Budget(max_agent_steps=100_000)
     n_calls = 500
     barrier = threading.Barrier(n_calls)
 
@@ -39,10 +39,10 @@ def test_concurrent_bump_tool_calls_count_correctly() -> None:
 
 def test_concurrent_bump_llm_usage_coherent() -> None:
     """N concurrent bump_llm_usage(...) calls must leave the three
-    counters in lockstep. Note: bump_llm_usage does NOT bump `turns`
+    counters in lockstep. Note: bump_llm_usage does NOT bump `agent_steps`
     — turn-counting is per agent step, not per LLM call (one step
     may issue N calls)."""
-    budget = Budget(max_turns=100_000)
+    budget = Budget(max_agent_steps=100_000)
     n_calls = 300
     barrier = threading.Barrier(n_calls)
 
@@ -53,34 +53,34 @@ def test_concurrent_bump_llm_usage_coherent() -> None:
     with ThreadPoolExecutor(max_workers=n_calls) as pool:
         list(pool.map(lambda _: worker(), range(n_calls)))
 
-    assert budget.turns == 0  # bump_llm_usage does not bump turns
+    assert budget.agent_steps == 0  # bump_llm_usage does not bump turns
     assert abs(budget.cost_usd - n_calls * 0.01) < 1e-6
     assert budget.prompt_tokens == n_calls
     assert budget.completion_tokens == 2 * n_calls
 
 
-def test_concurrent_bump_turn_count_correctly() -> None:
-    """N concurrent bump_turn() calls leave `turns == N`. Separate from
-    bump_llm_usage so `max_turns` caps agent steps, not LLM API calls."""
-    budget = Budget(max_turns=100_000)
+def test_concurrent_bump_agent_step_count_correctly() -> None:
+    """N concurrent bump_agent_step() calls leave `turns == N`. Separate from
+    bump_llm_usage so `max_agent_steps` caps agent steps, not LLM API calls."""
+    budget = Budget(max_agent_steps=100_000)
     n_calls = 300
     barrier = threading.Barrier(n_calls)
 
     def worker() -> None:
         barrier.wait()
-        budget.bump_turn()
+        budget.bump_agent_step()
 
     with ThreadPoolExecutor(max_workers=n_calls) as pool:
         list(pool.map(lambda _: worker(), range(n_calls)))
 
-    assert budget.turns == n_calls
+    assert budget.agent_steps == n_calls
     assert budget.cost_usd == 0.0
 
 
 def test_mixed_bumps_exhausted_check_coherent() -> None:
     """The `exhausted` read must be coherent against parallel bumps —
     no torn reads or misordered field checks."""
-    budget = Budget(max_turns=100_000, max_tool_calls=200)
+    budget = Budget(max_agent_steps=100_000, max_tool_calls=200)
     n_calls = 200
     barrier = threading.Barrier(n_calls)
     observed_exhausted = []
